// functions/src/types/PatientRelationshipType.ts
enum PatientRelationshipType {
  SAME_HOUSEHOLD = 'SAME_HOUSEHOLD',
  HUSBAND = 'HUSBAND',
  WIFE = 'WIFE',
  FATHER = 'FATHER',
  MOTHER = 'MOTHER',
  SON = 'SON',
  DAUGHTER = 'DAUGHTER',
  BROTHER = 'BROTHER',
  SISTER = 'SISTER',
  CONTACT = 'CONTACT',
  NEPHEW = 'NEPHEW',
  NIECE = 'NIECE',
  RELATIVE = 'RELATIVE',
  EXPOSURE = 'EXPOSURE',
}

export default PatientRelationshipType;